commit before refactoring axios requests

+++ routes/api/users.js
@@ -7,6 +7,7 @@ const passport = require('passport');
 // load input validation
 const validateRegisterInput = require('./../../validation/register');
 const validateLoginInput = require('./../../validation/login');
+const validateAccountEditInput = require('./../../validation/account');
 const isEmpty = require('./../../validation/is-empty');
 
 // load keys
@@ -79,7 +80,7 @@ router.post('/login', (req, res) => {
         .then(isMatch => {
           if (!isMatch) {
             errors.password = 'Password is incorrect';
-            return res.status(404).json(errors);
+            return res.status(400).json(errors);
           }
           const payload = { _id: foundUser._id, name: foundUser.name, isAdmin: foundUser.isAdmin };
           jwt.sign(payload, keys.secretOrKey, { expiresIn: 3600 }, (err, token) => {
@@ -98,6 +99,46 @@ router.post('/login', (req, res) => {
     .catch(err => res.status(400).json(err));
 });
 
+// @route     put /api/users/
+// @desc      edit user info
+// @access    private
+router.put('/', passport.authenticate('jwt', { session: false }), (req, res) => {
+  let errors = {};
+  User.findById(req.user._id).then(foundUser => {
+    bcrypt
+      .compare(req.body.oldPassword, foundUser.password)
+      .then(async isMatch => {
+        errors = validateAccountEditInput(req.body);
+        if (!isMatch) errors.oldPassword = 'Password is incorrect';
+        if (req.body.email) {
+          User.findOne({ email: req.body.email }).then(matchedUser => {
+            if (matchedUser && !matchedUser._id.equals(req.user._id))
+              errors.email = 'This email has already been taken';
+          });
+        }
+        if (!isEmpty(errors)) return res.json({ errors });
+        if (req.body.name) foundUser.name = req.body.name;
+        if (req.body.email) foundUser.email = req.body.email;
+        if (req.body.newPassword) {
+          await bcrypt
+            .genSalt(10)
+            .then(salt => {
+              if (!salt) throw 'Error generating salt.';
+              return bcrypt.hash(req.body.newPassword, salt);
+            })
+            .then(hash => {
+              if (!hash) throw 'Error generating hash.';
+              foundUser.password = hash;
+            })
+            .catch(err => res.status(400).json(err));
+        }
+        await foundUser.save();
+        res.json(foundUser);
+      })
+      .catch(err => res.json(err));
+  });
+});
+
 // @route     get /api/users/current
 // @desc      return current user info
 // @access    private
@@ -112,11 +153,22 @@ router.get('/current', passport.authenticate('jwt', { session: false }), (req, r
 // @route     delete /api/users
 // @desc      return current user info
 // @access    private
-router.delete('/', passport.authenticate('jwt', { session: false }), (req, res) => {
-  Profile.findOneAndRemove({ user: req.user._id })
-    .then(() => User.findByIdAndRemove(req.user._id))
-    .then(() => res.json({ success: true }))
-    .catch(err => res.status(404).json(err));
+router.post('/delete', passport.authenticate('jwt', { session: false }), (req, res) => {
+  const errors = {};
+  if (!req.body.password) {
+    errors.password = 'Please enter your password';
+    return res.json({ errors });
+  }
+  bcrypt.compare(req.body.password, req.user.password).then(isMatch => {
+    if (!isMatch) {
+      errors.password = 'Password is incorrect';
+      return res.json({ errors });
+    }
+    Profile.findOneAndRemove({ user: req.user._id })
+      .then(() => User.findByIdAndRemove(req.user._id))
+      .then(() => res.json({ success: true }))
+      .catch(err => res.status(404).json(err));
+  });
 });
 
 module.exports = router;
